requirements.txt
streamlit
llama-index
chromadb
cohere
python-dotenv
pandas
python-multipart
llama-index-core
llama-index-readers-file
llama-index-llms-cerebras
llama-index-embeddings-cohere
ipython
llama-index-experimental
llama-index-vector-stores-chroma
llama-index-llms-groq
llama-index-llms-gemini